refactor: consolidate general utility functions into `general-utils` and update references in tests
