+++ typescript/test/e2e/get-transaction-record.e2e.test.ts
@@ -6,7 +6,7 @@ import {
   type LangchainTestSetup,
 } from '../utils';
 import { Client, TransactionId } from '@hashgraph/sdk';
-import { extractObservationFromLangchainResponse, wait } from '../utils/general-util';
+import { extractObservationFromLangchainResponse, wait } from '../utils/general-utils';
 import Long from 'long';
 
 describe('Get Transaction Record E2E Tests', () => {
