+++ typescript/vitest.config.ts
@@ -31,5 +31,6 @@ export default defineConfig({
     },
     setupFiles: [],
     testTimeout: 60000,
+    hookTimeout: 60000,
   },
 });
